Move BuilderHelpers to antoher file and move some things around
