+++ DapperDoodle/DapperDoodle.csproj
@@ -2,7 +2,6 @@
 
     <PropertyGroup>
         <TargetFramework>netstandard2.1</TargetFramework>
-        <RootNamespace>DapperDiddle</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
